requirements.txt
pyrogram
pymongo
python-dotenv
transformers
torch
